+++ src/horton_part/gisa.py
@@ -115,7 +115,18 @@ class GaussianIterativeStockholderWPart(ISAWPart):
         return self.get_grid(index).rgrid
 
     def get_proatom_rho(self, iatom, propars=None):
-        """Get proatom density for atom `iatom`."""
+        """Get pro-atom density for atom `iatom`.
+
+        If `propars` is `None`, the cache values are used; otherwise, the `propars` are used.
+
+        Parameters
+        ----------
+        iatom: int
+            The index of atom `iatom`.
+        propars: np.array
+            The pro-atom parameters.
+
+        """
         if propars is None:
             propars = self.cache.load("propars")
         rgrid = self.get_rgrid(iatom)
@@ -124,6 +135,21 @@ class GaussianIterativeStockholderWPart(ISAWPart):
         return self.bs_helper.compute_proatom_dens(propars, alphas, rgrid.points, 1)
 
     def eval_proatom(self, index, output, grid):
+        """Evaluate function on a local grid.
+
+        The size of the local grid is specified by the radius of the sphere where the local grid is considered.
+        For example, when the radius is `np.inf`, the grid corresponds to the whole molecular grid.
+
+        Parameters
+        ----------
+        index: int
+            The index of an atom in the molecule.
+        output: np.array
+            The size of `output` should be the same as the size of the local grid.
+        grid: np.array
+            The local grid.
+
+        """
         propars = self.cache.load("propars")
         populations = propars[self._ranges[index] : self._ranges[index + 1]]
         exponents = self.bs_helper.load_exponent(self.numbers[index])
